+++ src/services/index.ts
@@ -1,4 +1,5 @@
 export * from './coding.ts';
+export * from './lruCache.ts';
 export * from './processEventSet.ts';
 export * from './systemRecord.ts';
 export * from './utils.ts';
